policy-employee-backtest-docker/BE/requirements.txt
pandas
python-dotenv
langchain
langchain-openai
fastapi
uvicorn